+++ editors/tree-sitter-jake/queries-src/highlights.scm
@@ -44,6 +44,10 @@
   path: (string) @string
   namespace: (identifier) @namespace)
 
+; Import rooted modifier
+(import_statement
+  "rooted" @keyword.control.import)
+
 ; Recipe attributes (directives above recipes)
 (recipe_attribute
   name: (identifier) @constant)
@@ -114,6 +118,7 @@
 (require_directive) @keyword.directive
 (export_directive) @keyword.directive
 (default_directive) @keyword.directive
+(rooted_directive) @keyword.directive
 (global_hook) @keyword.directive
 
 ; Import statement